Changed projects to groupings
Per discussions with @jinokian I've changed projects to groupings to better
inform the logical grouping of animals. This includes changes to how projects
are entered, viewed, and selected from dropdowns. Changes to the API as well.

I've also begun working on #41. The handling module is a small proof of concept.

Resolve: #40, #31