Fix data entry close #43. Close #44 as it pertains to API
Fixed all the issues concerning data entry. This includes fields that
are required or not. Added options to some dropdowns. Added an N/A and
Other to some dropdowns when the options isn't available. In the future
we may want to add an option to add to list and store these options
in a database.

References: #43, #44, #42
Closes #42 #43 $44